+++ packages/visualization-playground/jest.config.js
@@ -23,36 +23,6 @@
  SOFTWARE.
 
  */
-import type { FC } from 'react'
-import React, { useState, useEffect } from 'react'
-import { FieldCheckbox } from '@looker/components'
+const { config } = require('@looker/jest-config')
 
-interface CheckboxProps {
-  label?: string
-  onChange?: (isChecked: boolean) => void
-  checked?: boolean
-}
-
-/**
- * A simple checkbox designed to be symmetrical with other components used in
- * Config editor. Stores checked state, and accepts a "checked" prop (rather
- * than "checked" per standard checkbox)
- */
-
-export const Checkbox: FC<CheckboxProps> = ({ onChange, label, checked }) => {
-  const [isChecked, setIsChecked] = useState(Boolean(checked))
-  useEffect(() => {
-    if (checked !== isChecked) {
-      onChange?.(isChecked)
-    }
-  }, [isChecked, onChange, checked])
-  return (
-    <FieldCheckbox
-      label={label}
-      checked={isChecked}
-      onChange={() => {
-        setIsChecked(!isChecked)
-      }}
-    />
-  )
-}
+module.exports = config