+++ packages/visualization-playground/CHANGELOG.md
@@ -1,7 +1,48 @@
-# Change Log
+# [2.1.0](https://github.com/looker-open-source/components/compare/22.16.0...2.1.0) (2022-10-12)
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+### Features
+
+* add show_totals and show_row_totals config transformation b/248616608 ([8fda5e1](https://github.com/looker-open-source/components/commit/8fda5e15d6622b9928399f8e13393b3a8383973c))
+
+
+
+# [2.0.0](https://github.com/looker-open-source/components/compare/22.16.0...2.0.0) (2022-09-22)
+
+### Features
+* Toggle config.show_totals for table visualizations
+* Render updated table visualization 
+
+# [1.4.0](https://github.com/looker-open-source/components/compare/22.10.3...1.4.0) (2022-09-07)
+
+### Features
+
+* Add ability to toggle config.truncate_header for table visualizations
+
+### Reverts
+
+* Revert "Move no-restricted-imports to eslint plugin" ([3c5d40b](https://github.com/looker-open-source/components/commit/3c5d40b14453034528a39a6d52261c64efd43082))
+
+
+
+# [1.3.0](https://github.com/looker-open-source/components/compare/22.10.3...1.3.0) (2022-07-29)
+
+Adds theme provider within package
+
+### Reverts
+
+* Revert "Move no-restricted-imports to eslint plugin" ([3c5d40b](https://github.com/looker-open-source/components/commit/3c5d40b14453034528a39a6d52261c64efd43082))
+
+
+
+# [1.2.0](https://github.com/looker-open-source/components/compare/22.10.3...1.2.0) (2022-07-29)
+
+Update package json `main` setting
+
+## [1.1.0](https://github.com/looker-open-source/components/compare/22.10.3...1.1.0) (2022-07-28)
+
+Update looker-components peer dependency.
+
 
 ## [1.0.4](https://github.com/looker-open-source/components/compare/@looker/visualization-playground@1.0.3...@looker/visualization-playground@1.0.4) (2022-05-23)
 
@@ -30,9 +71,3 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 1.0.1 (2022-04-29)
 
 **Note:** Version bump only for package @looker/visualization-playground
-
-
-
-
-
-# Change Log